+++ api/index.js
@@ -5,9 +5,6 @@ const faker = require('faker');
 const isEmpty = require('lodash/isEmpty');
 
 const NB_GENERATED_ENTITY = 10;
-
-faker.locale = "fr";
-
 let entityJSON = [{
     uuid: '1234',
     firstName: 'Don Diego',
@@ -67,9 +64,9 @@ for(let i = 0; i < NB_GENERATED_ENTITY; i++){
 }
 
 function _createFinancialMoves(){
-    return [0,1,2,3,4,5].map(() => ({
+    return [0,1,2,3,4,5,6,7,8,9,10].map(() => ({
         transactionType: faker.finance.transactionType(),
-        amount: faker.finance.amount()
+        amount:faker.finance.amount()
     }));
 }
 
@@ -85,15 +82,15 @@ const MOCKED_API_PORT = process.env.API_PORT || 9999;
 /*****************************************
 ************** Mocked API ****************
 ******************************************/
+
 const API_ROOT = '/x';
 const app = express();
-
 //middleware
+//app.use(express.static(staticFolder));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({
     extended: true
 }));
-
 app.use((req, res, next) => {
     console.log(new Date() + ', ' + req.method + ', ' + req.url);
     if (!isEmpty(req.body)) {
@@ -101,7 +98,6 @@ app.use((req, res, next) => {
     }
     next();
 });
-
 //CORS middleware
 const allowCrossDomain = (req, res, next) => {
     res.header('Access-Control-Allow-Origin', '*');
@@ -110,7 +106,6 @@ const allowCrossDomain = (req, res, next) => {
     res.header('Content-Type', 'application/json');
     next();
 }
-
 app.use(allowCrossDomain);
 
 app.get(API_ROOT + '/entity', function getAllNotifications(req, res) {
@@ -142,14 +137,8 @@ app.get(API_ROOT + '/finances/:id', function getSingleFinanceEntity(req, res) {
     res.json({finance: entityFinanceJSON.find(d => d.uuid === req.params.id)});
 });
 
-app.get(API_ROOT + '/complex', (req, res) => {
+app.get(API_ROOT  + '/complex', (req, res) => {
     res.json(complexJSON);
-});
-
-app.put(API_ROOT + '/entity/:id', (req, res) => {
-    var savedData = req.body;
-    savedData.isSaved = true;
-    res.json(savedData);
 });
 
 app.post(API_ROOT + '/entity/:id', (req, res) => {
