+++ src/config/domains/index.js
@@ -1,12 +1,12 @@
 import React, {Component, PropTypes} from 'react';
 
 import Checkbox from 'focus-components/input-checkbox';
+import Radio from 'focus-components/input-radio';
 import InputDate from 'focus-components/input-date';
 import InputText from 'focus-components/input-text';
 import InputSelect from 'focus-components/select-mdl';
 import RadioSelect from 'focus-components/select-radio';
-import Autocomplete from 'focus-components/autocomplete-text/field';
-import SelectCheckbox from 'focus-components/select-checkbox';
+import Autocomplete from 'focus-components/autocomplete-text/edit';
 
 export const DO_ID = {
     type: 'text'
@@ -50,15 +50,35 @@ export const DO_SEXE = {
 
 export const DO_ACCOUNTS_NAMES = {
     type: 'text',
-    InputComponent: props => <div><Autocomplete/></div>
+   InputComponent: props => <div><Autocomplete querySearcher={_querySearcher}/></div>
 }
 
 export const DO_CHECKBOX = {
-    type: 'boolean',
+    type: 'checkbox',
     InputComponent: Checkbox
 }
 
-export const DO_SELECT_CHECKBOX = {
-    type: 'boolean',
-    InputComponent: SelectCheckbox
-}
+const _querySearcher = query => {
+    let data = [
+        {
+            key: 'JL',
+            label: 'Joh Lickeur'
+        },
+        {
+            key: 'GK',
+            label: 'Guénolé Kikabou'
+        },
+        {
+            key: 'YL',
+            label: 'Yannick Lounivis'
+        }
+    ];
+    return new Promise((resolve, reject) => {
+        setTimeout(() => {
+            resolve({
+                data,
+                totalCount: data.length
+            });
+        }, 500);
+    });
+};
