+++ src/containers/dev-tools.js
@@ -13,7 +13,7 @@ const DevTools = createDevTools(
     // Consult their repositories to learn about those props.
     // Here, we put LogMonitor inside a DockMonitor.
     <DockMonitor toggleVisibilityKey='ctrl-h' changePositionKey='ctrl-q' defaultSize={0.4} defaultIsVisible={false}>
-            <Inspector />
+        <Inspector />
     </DockMonitor>
 );
